requirements.txt
discord.py>=2.3.0,<3.0.0
jishaku>=2.5.1,<3.0.0
